pc - set up communication between front end and back end
This is a preliminary step to fetching and retrieving the actual course data.
For now, the backend route /api/public/basicsearch just takes the values
of the three request params (qtr,dept,level) and echoes them back in a JSON
object.

Eventually, this route will perform the query and return the results.