pc - refactor JSONPrettyPanel by renaming and fixing nesting error
JSONPrettyPanel was previous based on a Panel component
but is now based on a Card component, so we renamed it.

Also, since JSONPretty contains a div element, it cannot
appear inside Card.Text which formats a p element.  So
we removed the Card.Text around it and placed it directly
inside Card.Body.