pc - make course history search use MongoDB
Plus, partial test coverage of backend.
We had to create a different search form
because we were not able to find a field that corresponded to
level in the MongoDB database.  If we want to implement that
search, we may need a different strategy.

Some unfinished work here is test coverage for the
new controller, and for the modifications to the front end.