pc - make default subjectArea first one in hardcoded fixture
There is a hardcoded fixture which is a snapshot of the list
of subject areas; we fall back to that if/when there are API
problems.   We can also use that as a source of the likely
first element that will come from the API call.  This needs to match
otherwise, a query that comes before that will result in the wrong
course info being retrieved.