pc - add swagger-ui endpoint and /api/docs endpoints for RESTful API doc and testing
This adds an endpoint /swagger-ui/ with a UI for documenting and testing the RESTful API endpoints
plus a JSON endpoint /api/docs, both available without authentication.

Note that authenticated endpoints still require a JWT.   One can be obtained if
and only if you are already logged in, by copying/pasting from the developer tools
and pasting into the Authentication field.